requirements.txt
dash
dash-bootstrap-components
psycopg2-binary
